Frontend Updates

+++ modules/database_modules/class_database.py
@@ -49,6 +49,28 @@ class ClassesDatabase:
             try:
                 cur = conn.cursor()
 
+                # Check if the teacher exists
+                check_query = """
+                    SELECT 1 FROM users_teachers
+                    WHERE id = %s;
+                """
+                cur.execute(check_query, (filtered_kwargs['teacher_id'],))
+                existing_teacher = cur.fetchone()
+                if not existing_teacher:
+                    return self.generate_response(success=False, error='Teacher not found.', status_code=404)
+
+                # Check if the class already exists
+                check_query = """
+                    SELECT 1 FROM classes
+                    WHERE class_name = %s AND teacher_id = %s AND group_num = %s AND semester = %s;
+                """
+                cur.execute(check_query,
+                            (filtered_kwargs['class_name'], filtered_kwargs['teacher_id'],
+                             filtered_kwargs['group_num'], filtered_kwargs['semester']))
+                existing_class = cur.fetchone()
+                if existing_class:
+                    return self.generate_response(success=False, error='Class already exists.', status_code=400)
+
                 # Insert the class data into the database
                 columns = ', '.join(filtered_kwargs.keys())
                 values = ', '.join([f'%({field})s' for field in filtered_kwargs.keys()])
@@ -72,26 +94,40 @@ class ClassesDatabase:
                 print(f"Error registering class: {error_message}") # Debugging print
                 return self.generate_response(success=False, error=error_message, status_code=500, error_code=e.pgcode)
 
-    def retrieve_classes(self, teacher_id):
-        if not teacher_id:
-            return self.generate_response(success=False, error='Teacher ID must be provided.', status_code=400)
+    def retrieve_class_students(self, class_id):
+        if not class_id:
+            return self.generate_response(success=False, error='Class ID must be provided.', status_code=400)
 
         with db_connection(self.credentials) as conn:
             try:
                 cur = conn.cursor(cursor_factory=psycopg2.extras.DictCursor)
-                query = """
-                    SELECT * FROM classes
-                    WHERE teacher_id = %s;
+
+                # Check if the class exists
+                check_query = """
+                    SELECT 1 FROM classes
+                    WHERE class_id = %s;
                 """
-                cur.execute(query, (teacher_id,))
-                classes = cur.fetchall()
+                cur.execute(check_query, (class_id,))
+                existing_class = cur.fetchone()
+                if not existing_class:
+                    return self.generate_response(success=False, error='Class not found.', status_code=404)
+
+                # Retrieve the students registered in the class
+                query = """
+                    SELECT users_students.id, name, username, email, faculty, matnum
+                    FROM users_students
+                    JOIN classes_students ON users_students.id = classes_students.student_id
+                    WHERE class_id = %s;
+                """
+                cur.execute(query, (class_id,))
+                students = cur.fetchall()
                 cur.close()
-                classes_dict = [dict(row) for row in classes]
-                return self.generate_response(success=True, error=None, status_code=200, data=classes_dict)
+                students_dict = [dict(row) for row in students]
+                return self.generate_response(success=True, error=None, status_code=200, data=students_dict)
 
             except psycopg2.Error as e:
                 error_message = e.pgerror if e.pgerror else str(e)
-                print(f"Error retrieving classes: {error_message}")
+                print(f"Error retrieving class students: {error_message}")
                 return self.generate_response(success=False, error=error_message, status_code=500, error_code=e.pgcode)
 
     def update_class(self, class_id, **kwargs):
@@ -111,6 +147,16 @@ class ClassesDatabase:
         with db_connection(self.credentials) as conn:
             try:
                 cur = conn.cursor()
+
+                # Check if the class exists
+                check_query = """
+                    SELECT 1 FROM classes
+                    WHERE class_id = %s;
+                """
+                cur.execute(check_query, (class_id,))
+                existing_class = cur.fetchone()
+                if not existing_class:
+                    return self.generate_response(success=False, error='Class not found.', status_code=404)
 
                 # Construct the query dynamically
                 set_clause = ', '.join([f'{field} = %({field})s' for field in filtered_kwargs.keys()])
@@ -143,6 +189,18 @@ class ClassesDatabase:
         with db_connection(self.credentials) as conn:
             try:
                 cur = conn.cursor()
+
+                # Check if the class exists
+                check_query = """
+                    SELECT 1 FROM classes
+                    WHERE class_id = %s;
+                """
+                cur.execute(check_query, (class_id,))
+                existing_class = cur.fetchone()
+                if not existing_class:
+                    return self.generate_response(success=False, error='Class not found.', status_code=404)
+
+                # Delete the class
                 query = """
                     DELETE FROM classes
                     WHERE class_id = %s
@@ -160,24 +218,47 @@ class ClassesDatabase:
                 print(f"Error deleting class: {error_message}")
                 return self.generate_response(success=False, error=error_message, status_code=500, error_code=e.pgcode)
 
-    def add_student_to_class(self, student_id, class_id):
-        if not student_id or not class_id:
-            return self.generate_response(success=False, error='Both student ID and class ID must be provided.', status_code=400)
+    def add_student_to_class(self, matnum, class_id):
+        if not matnum or not class_id:
+            return self.generate_response(success=False, error='Both matnum and class ID must be provided.', status_code=400)
 
         with db_connection(self.credentials) as conn:
             try:
                 cur = conn.cursor()
 
+                # Check if the class exists
+                check_query = """
+                    SELECT 1 FROM classes
+                    WHERE class_id = %s;
+                """
+                cur.execute(check_query, (class_id,))
+                existing_class = cur.fetchone()
+                if not existing_class:
+                    return self.generate_response(success=False, error='Class not found.', status_code=404)
+
+                # Check if the student exists and retrieve the student ID
+                check_query = """
+                    SELECT id FROM users_students
+                    WHERE matnum = %s;
+                """
+                cur.execute(check_query, (str(matnum),))  # Convert matnum to string
+                existing_student = cur.fetchone()
+                if not existing_student:
+                    return self.generate_response(success=False, error='Student not found.', status_code=404)
+
+                student_id = existing_student[0]
+
                 # Check if the student is already registered in the class
                 query = """
-                    SELECT * FROM classes_students
+                    SELECT 1 FROM classes_students
                     WHERE student_id = %s AND class_id = %s;
                 """
                 cur.execute(query, (student_id, class_id))
-                existing_student = cur.fetchone()
-                if existing_student:
+                existing_registration = cur.fetchone()
+                if existing_registration:
                     return self.generate_response(success=False, error='Student is already registered in the class.', status_code=400)
 
+                # Insert the student into the class
                 query = """
                     INSERT INTO classes_students (student_id, class_id)
                     VALUES (%s, %s)
@@ -194,30 +275,6 @@ class ClassesDatabase:
                 print(f"Error adding student to class: {error_message}")
                 return self.generate_response(success=False, error=error_message, status_code=500, error_code=e.pgcode)
 
-    def retrieve_class_students(self, class_id):
-        if not class_id:
-            return self.generate_response(success=False, error='Class ID must be provided.', status_code=400)
-
-        with db_connection(self.credentials) as conn:
-            try:
-                cur = conn.cursor(cursor_factory=psycopg2.extras.DictCursor)
-                query = """
-                    SELECT users_students.id, name, username, email, faculty, matnum
-                    FROM users_students
-                    JOIN classes_students ON users_students.id = classes_students.student_id
-                    WHERE class_id = %s;
-                """
-                cur.execute(query, (class_id,))
-                students = cur.fetchall()
-                cur.close()
-                students_dict = [dict(row) for row in students]
-                return self.generate_response(success=True, error=None, status_code=200, data=students_dict)
-
-            except psycopg2.Error as e:
-                error_message = e.pgerror if e.pgerror else str(e)
-                print(f"Error retrieving class students: {error_message}")
-                return self.generate_response(success=False, error=error_message, status_code=500, error_code=e.pgcode)
-
     def del_student_from_class(self, student_id, class_id):
         if not student_id or not class_id:
             return self.generate_response(success=False, error='Both student ID and class ID must be provided.', status_code=400)
@@ -225,6 +282,26 @@ class ClassesDatabase:
         with db_connection(self.credentials) as conn:
             try:
                 cur = conn.cursor()
+
+                # Check if the class exists
+                check_query = """
+                    SELECT 1 FROM classes
+                    WHERE class_id = %s;
+                """
+                cur.execute(check_query, (class_id,))
+                existing_class = cur.fetchone()
+                if not existing_class:
+                    return self.generate_response(success=False, error='Class not found.', status_code=404)
+
+                # Check if the student exists
+                check_query = """
+                    SELECT 1 FROM users_students
+                    WHERE id = %s;
+                """
+                cur.execute(check_query, (student_id,))
+                existing_student = cur.fetchone()
+                if not existing_student:
+                    return self.generate_response(success=False, error='Student not found.', status_code=404)
 
                 # Check if the student is registered in the class
                 check_query = """
@@ -263,6 +340,18 @@ class ClassesDatabase:
         with db_connection(self.credentials) as conn:
             try:
                 cur = conn.cursor(cursor_factory=psycopg2.extras.DictCursor)
+
+                # Check if the class exists
+                check_query = """
+                    SELECT 1 FROM classes
+                    WHERE class_id = %s;
+                """
+                cur.execute(check_query, (class_id,))
+                existing_class = cur.fetchone()
+                if not existing_class:
+                    return self.generate_response(success=False, error='Class not found.', status_code=404)
+
+                # Retrieve the exams for the class
                 query = """
                     SELECT * FROM exams
                     WHERE class_id = %s;
@@ -288,7 +377,6 @@ class ClassesDatabase:
                 error_message = e.pgerror if e.pgerror else str(e)
                 print(f"Error retrieving class exams: {error_message}")
                 return self.generate_response(success=False, error=error_message, status_code=500, error_code=e.pgcode)
-
 
     # Private method to generate a consistent JSON response
     @staticmethod
